feat: Merged VL's Prefab naiively into PrefabManager. TODO: Integrate tightly with abstractions.
